Move getAutoScalingGroupDesiredCapacity function to autoscaling.go
