+++ myaws/autoscaling_waiter.go
@@ -7,7 +7,6 @@ import (
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/aws/request"
 	"github.com/aws/aws-sdk-go/service/autoscaling"
-	"github.com/pkg/errors"
 )
 
 // WaitUntilAutoScalingGroupStable is a helper function which waits until
@@ -87,23 +86,6 @@ func (client *Client) waitUntilAutoScalingGroupNumberOfInstancesEqualsDesiredCap
 	w.ApplyOptions(opts...)
 
 	return w.WaitWithContext(ctx)
-}
-
-// getAutoScalingGroupDesiredCapacity is a helper function which returns
-// DesiredCapacity of the specific AutoScalingGroup.
-func (client *Client) getAutoScalingGroupDesiredCapacity(asgName string) (int64, error) {
-	input := &autoscaling.DescribeAutoScalingGroupsInput{
-		AutoScalingGroupNames: []*string{&asgName},
-	}
-
-	response, err := client.AutoScaling.DescribeAutoScalingGroups(input)
-	if err != nil {
-		return 0, errors.Wrap(err, "getAutoScalingGroupDesiredCapacity failed:")
-	}
-
-	desiredCapacity := response.AutoScalingGroups[0].DesiredCapacity
-
-	return *desiredCapacity, nil
 }
 
 // waitUntilAutoScalingGroupAllInstancesAreInService waits until all instances
